Smarter logic to auto-detect the right attributes.
This will walk back the superclass chain looking for an attribute that
is set. This means when you do a subclass of an inversion provider in
its own cookbook (ex. poise-service-runit, poise-ruby-build) it will
still try to use the base cookbook's name for the attribute. Not
foolproof, but should be enough in most cases. Also means we get a
nicer error message if the default isn't available at all.