Major overhaul of the options resource mixin.
This requires changing the run_state data format to include the resource
type. Previously two resources of different types with the same name
would have crossed the streams. Also adds support for using an option
key as an rvalue once it is defined and adds tests for the options
resource mixin.

Thanks to @soosiechoi for her support on Kickstarter!