Updates for Chef 12.4 to match @jkeiser's changes.
All of this should be backwards compatible with earlier Chef. Sets the
actual resource name as the Chef-level #provides() on inversion
providers so they get hooked up correctly to their resource. Switches
the subresource container code to not use method_missing. This
simplifies the handling of created_at a bit. Removing the use of
method_missing might also allow removing the to_ary trick in Chef 13.
The inversion provider unit test was only ever working by accident,
I never actually set the inversion provider to use, using provides_auto
for now so it will be the default. Also changed the provider class
name to catch this in the future if there is a reversion.