Don't apply my implementation of lazy defaults on 12.5+.
Ping @jkeiser for sanity check. Looks like this landed in master
recently but should be directly compatible so I can just disable
my code based on the Chef version. Includes a test for mutable, lazy
defaults because before this change, the upstream code in _pv_default
wouldn't see it was lazy and would freeze the default value.